Enable event finishing (#360)
This PR introduces a new feature for the async mode:
`/adept/FinishLastNParticlesOnCPU`.

It is not complete yet, as they are currently just killed and not pushed
to the CPU, as this will require #358 to work and be merged.

Still, killing the last N particles can currently be used to assess some
performance bottlenecks with the magnetic field.

Also, a timer is added to the printout of the async printouts.

This allows for the plotting the numbers of flight vs time and not just
only vs iterations, as this severely skews it towards the fast
iterations at the tail:
<img width="637" alt="Screenshot 2025-03-16 at 10 45 45"
src="https://github.com/user-attachments/assets/b09065a3-567f-4e92-8528-95d60ff3b772"
/>