enable full uniform B field via RK integrator (#368)
Previously, only a constant Bz field was available.

As needed for the B5 example, a full uniform magnetic field is required.
There was an not-up-to-date analytical Bany field integrator. Since it
was not up to date, here I propose to simply use the RK integrator.

Now, similar to the GeneralMagneticField, a UniformMagneticField is
created on the host, and passed to the device.

This cleans up quite a bit of logic on the B field.