Improve Geant4 version test (#383)
Shamelessly stolen from @drbenmorgan's [PR to G4HepEm
](https://github.com/mnovak42/g4hepem/pull/125), this improves the
version test in AdePT: instead of testing for version > 11, now it is
tested for the `G4VTrackingManager`, as this one is required. This fixes
a compilation issue in Gauss which uses a patched version of 10.7 with
the G4VTrackingManager.

The old test for flush which was needed for the FastSim hook approach is
removed.