Update validation reference data and use lower statistics for test (#384)
Thanks to https://github.com/apt-sim/AdePT/pull/381 our agreement with
G4HepEM is a factor of 10 better. This means that we can afford to use
lower statistics in our validation test for a faster CI.

This PR updates the reference data taken from a pure CPU G4HepEM run,
and reduces the amount of events used in the validation test.

**Validation plot shooting 1e4 particles** (The new test run)

![plot_1e4](https://github.com/user-attachments/assets/34218253-51b3-43a7-84f3-f7cfe38b79a5)

**Validation plot shooting 1e5 particles**

![plot](https://github.com/user-attachments/assets/f6b48af5-6ea6-4599-ac95-bf5dac862873)