Initialize AdePT from G4 workers (#386)
In the experimental frameworks, AdePT must be initialized by the G4
workers, because at the initialization time of the master thread, the
number of G4 workers is not yet known.

Therefore, the initialization is postponed to the first G4 worker. Then,
all other G4 workers must wait for the first worker to have initialized
AdePT before they can proceed.

This fixes running with AdePT in Gauss and Athena.