Fix leaks in asynchronous mode (#358)
This fixes the extraction of leaked tracks in Async mode. The previous
extraction of leaks didn't work in most cases for two main reasons:
- The size of the buffer used for extracting the leaked tracks is
limited, so if there are too many, the extraction needs to be done in
multiple steps. However, there was no mechanism to do this, tracks would
get postponed until the next extraction step, but the corresponding
events were notified that all leaks had been extracted. This means that
whenever the number of leaks exceeded the buffer size the overflow was
lost.
- There wasn't a way to guarantee that all leaks corresponding to the
event requesting a flush had been extracted. The tracks are not
extracted in order of insertion, and the transport kernels were allowed
to insert new tracks into the leak queues in-between transfers of tracks
from device to host. This means that we couldn't know when all leaks
corresponding to an event had actually been extracted.

With these changes:
- The leak queues are frozen whenever an event requests a flush of their
leaks. We keep two leak queues per particle type, we swap them at the
start of the leak extraction, and don't swap them back until the
extraction ends and another event requests a flush. This ensures that
when we finish transferring the frozen queue, all leaks associated to
that event have been extracted.
- We use a new state machine that doesn't return to the idle state until
all leaks have been transferred. The functionality is similar to that of
the scoring thread, but instead of running separately from the transport
loop, the work is done by the AdePT thread.
- Returning all leaks uncovers an issue with the MPMC queues used to
store leaked tracks during transport. These queues have a fixed size and
need to be emptied before they can be used again, and it is left to the
user to check whether the queue is full. The queues were inadvertedly
overflowing in specific use cases (Such as the testEm3 geometry when
running large events), this has been solved by stopping transport in
situations where the queues are close to overflowing but a leak
extraction is already in progress. Since this is not an issue in more
realistic use cases, this mechanism ensures that the execution will be
error-free in all situations without affecting runtime in realistic
runs.
- In specific use cases like running large events in testEm3 the results
were sometimes non-reproducible. This was likely due to some
interference between the leak extraction, hit extraction, and particle
injection, which ran on the same stream. The issue is solved by running
these three operations on different streams, but should be revisited in
the future.

The new `ExtractStates` are the following:
- **Idle**: No extraction is in progress
- **ExtractionRequested**: An event requested a flush, waiting for
transport to finish
- We move from `Idle` to `ExtractionRequested` when an event has
requested a flush, and their `EventState` is `HitsFlushed`. The number
of particles in flight for that event is guaranteed to be 0 at this
point
- We wait for the transport kernels to finish running as they are
writing to the same leak queues that we need to transfer
- **TracksNeedTransfer**: An event requested a flush, leak buffer on
device has tracks to transfer
- Transport has finished, the leak queues are frozen and we can start
transferring tracks
-  **PreparingTracks**: Tracks are being copied to the staging buffer
- Tracks are copied from the leak queues to the transfer buffer, the
numbers of transferred and remaining tracks are copied to the host.
-  **TracksReadyToCopy**: Staging buffer is ready to be copied to host
-  **CopyingTracks**: Tracks are being copied to host
- **TracksOnHost**: Some or all the tracks have been transferred from
device to host and are waiting in the copy buffer
-  **SavingTracks**: Tracks are being copied to per-event queues
- Tracks are copied from the transfer buffer, which can contain tracks
for all events, into per-event queues
- **TracksSaved**: Tracks have been moved from the copy buffer to their
respective per-event queues
- All tracks are stored in the per-event queues, so the transfer buffer
can be re-used. If there are any remaining leaks on GPU the state goes
back to `TracksNeedTransfer`, otherwise the state moves to `Idle`


![extract_state_machine](https://github.com/user-attachments/assets/9d6ad298-a14d-416a-9a62-7f40838d27eb)