Separating slot manager into per-particle managers (#388)
Currently async mode uses one slotmanager, and one array for tracks, the
slotmanager keeps track of the usage of the track array, handing out
free slots and doing cleanup when necessary, and we keep per-particle
lists of slots occupied by that particle type. With this change, each
particle type has it's own slotmanager and array of tracks.

This is needed in order to implement the split kernels in async mode,
since in that case, in addition to the AdePT tracks we also keep G4HepEM
tracks, which need to be per-particle (This means that we cannot keep
all hepem tracks in the same array as we do with AdePT tracks)

In addition to that, this will allow for better coalesced memory
operations when using split kernels, as the loaded tracks will be more
contiguous in memory.