Fix global time in scoring (#392)
This PR fixes the global time in the scoring, as it is now passed back
to Geant4. Also, a bug is fixed: for calculating the global time, the
initial velocity must be used, before updating the kinetic energy due to
energy loss. Before, stopped particles would then lead to an infinite
global time. Now it conforms with G4.

The globalTime for secondaries is also corrected: before it was taking
the globaltime from `currentTrack`, however, the member
`currentTrack.globalTime` was only updated in survive, so after the
secondary was created. Now the current globalTime is passed to the
constructor.

Now, the time seems to be fully correct in the Gauss histograms