Split kernels in Async mode (#395)
Adding split kernels for the Async mode

- Replaces the old split kernels, which were outdated and worked only in
synchronous mode
- The new kernels are only available in async mode
- Enables testing of the split kernels on the CI

The new kernel split is:
- Gammas
  - GammaPropagation
  - GammaRelocation
- GammaInteractions (Optional between a single kernel or split
interaction kernels)
  - GammaConversion
  - GammaCompton
  - GammaPhotoelectric
- Electrons/Positrons:
  - ElectronHowFar
  - ElectronPropagation
  - ElectronMSC
  - ElectronRelocation
- ElectronInteractions (Optional between a single kernel or split
interaction kernels)
  - ElectronIonization
  - ElectronBremsstrahlung
  - PositronAnnihilation
  - PositronStoppedAnnihilation

The interactions are split through the use of per-interaction queues
where the indices of tracks performing that interaction are stored. In
order to avoid synchronizations and callbacks, the grid configuration is
not adjusted depending on the number of particles performing a
particular interaction. This means that part of the threads in a block
will be scheduled but inmediately become free as they will have no
tracks to transport. This could be improved in the future, but using
smaller grids doesn't seem to have any effect on performance.

The split kernels do not work yet with a B field since the cut loopers
are not taken into account and are transported in the following kernels
even if they were killed before.