fix synchronization issue between SwapDeviceBuffers and TransferHitsToHost (#396)
This PR fixes two issues, where the first one caused the CI to fail
frequently.

1. There was a race condition in `SwapDeviceBuffers` and
`TransferHitsToHost`:
The Transport Thread would call `SwapDeviceBuffers` and set the
DeviceState to `NeedTransferToHost` immediately, although the copy of
the numbers of GPU steps was copied asynchronously. Then, the Step
processing thread in `TransferHitToHost` would check for the device
state to be in the state `NeedTransferToHost` to start the transfer.
However, in rare occasions, the copy of the number of GPU steps to be
copied was not yet complete, leading to the wrong number of steps being
copied, which then subsequently led to threads processing the steps of a
different thread, which would throw an error and break reproducibility.
This is fixed with a `cudaLaunchHostFunc` call that sets the state only
when the transfer has finished.

2. In the hit management, the Transport thread was advancing the state 
```c++
            AdvanceEventStates(EventState::RequestHitFlush, EventState::FlushingHits, eventStates);
...
            gpuState.fHitScoring->SwapDeviceBuffers(gpuState.stream);
            hitProcessing->cv.notify_one();
```
Before the hitprocessing thread is notified.

But then the step management thread advances the state from
`FlushingHits` to `HitsFlushed`:
```c++
    if (haveNewHits) {
      AdvanceEventStates(EventState::FlushingHits, EventState::HitsFlushed, eventStates);
    }
```
As the transport and step management thread are detached, in principle,
the transport thread could advance the state to `FlushingHits` and the
the step management thread could just have finished the previous
`TransferHitsToHost` call and immediately advance the state to
`HitsFlushed`, although the hits have not yet been flushed. Then, those
steps could end up in the next event, breaking again reproducibility.
Note that case 2 was not observed but it was unsafe. The solution is
that only the step management thread is allowed to advance the state,
fixing this potential gap.

This PR might have fixed all issues that were temporary fixed in #389,
but I would be in favor of checking that separately with the reproducer
that was used for #389 before. In my current setup (which was derived
from the crashing CI), I could revert #389 on top of this PR and not
observe any errors anymore, but this should be confirmed and double
checked with the previous reproducer before reverting #389.