Fix synchronization between SwapDeviceBuffers and Transport (#401)
This PR fixes a subtle but important synchronization that was missing
between the swap of the device buffers and the transport.

In rare cases, this could lead to the following scenario:

The hit slot counter statistics were copied from the GPU to the CPU, and
reset to 0 on the device. Then the swap was executed.

However, after resetting the slot counter statistics and **before** the
swap was executed, the transport already ran again and wrote some steps
(then back to the initial position, as the counter was already reset),
so some later steps overwrote some initial steps in the buffer.

This race condition broke reproducibility and could result in wrong
results.

So far, no more issues with reproducibility are observed, so the CI test
is put back in place.