Split relocation kernels (#408)
Separate the Relocation kernels for Electrons and gammas into a _setup
kernel_ which decides whether a track will relocate, do an interaction,
or continue to the next step, and a _relocation kernel_ in which the
actual geometry calls are performed.

The reason for this split is that the current relocation kernels,
specially in the case of electrons and positrons are very divergent.
Only a small portion of the tracks in a warp will do a relocation, but
the geometry calls involved are costly, which can slow down the rest of
the threads in that warp.

These changes haven't shown a speedup so far in runs without a B-field,
but the split should improve profiling results by isolating the
relocation code.