Fix HitSlot overflow (#411)
This PR fixes the crash that happens if AdePT runs out of hit slots for
async AdePT in almost all cases.

It is checked whether there are more particles in flight than there are
hit slots left (as the number is only an estimate, I added a factor of
1.5). If that is the case, the next step might cause a crash. Then, one
has to wait for the DeviceBuffers to be ready to swapped, before one can
continue with the swap (which frees the hit slots), before the next step
can be taken.

Note that if more tracks are in flight than total hit slots are
available, any time all tracks could generate a hit and cause a crash
and overflow. This is not fixable by waiting for freeing the hit slots,
this requires simply more slots or less threads (which leads to more
slots per thread and less particles in flight).