Integrate in safety steps (#420)
The field propagator was limiting arc integration by the bending error
only, validating the advance against the current safety. This actually
uses the current safety to limit the integration step to a possibly
larger value.

We have to measure better, but I see a runtime gain of ~20% in CMS using
this