Fix CPU B field in examples, update covfie version (#425)
This fixes a crash that was observed running the example1 with an
external B field map:

While the B field on the GPU in texture memory is always clamped and
therefore secure against out-of-bounds access, the B field on the CPU
that we use in our examples is not. This requires initializing the B
field type as clamped.

As the B field type of the reader and writer must be the same in covfie,
(and we require an unclamped version for the texture memory), the
clamped B field is generated separately from the unclamped one, which is
read from file.

The covfie version is updated.