Add stricter compiler flags (#426)
This PR adds stricter compiler flags to AdePT. Additionally, a new
compile time option is introduced to enforce some of the flags. This
setting enforces the same error flags as CMSSW. If they are not
enforced, still, they are printed as warnings.

This new setting revealed many minor issues, but also a bigger one,
where the `parentTrackInfo` was used instead of the `hostTrackInfo`.

Note that even stricter flags could be set, but those reveal countless
issues in VecGeom and G4VG, so it is left for future cleaning.