Store leaked traks in a separate array (#424)
Currently, there is no dedicated buffer for leaked tracks, they are
stored directly on the array of tracks being transported.

This PR adds a separate buffer for leaks, which implies copying each
leaked track to it. The advantage of this approach is that we can
inmediately mark the slot for freeing on the active tracks array.

The main reason for this PR is that storing the leaked tracks along with
the others makes it very hard to defragment the track array as would be
needed if track data was stored in an SoA layout.

This change doesn't affect the run times.

`RTX4080, CMS2026 8 threads x 32 events, 3.8T Bfield`
| Kernels    | Time Master | New Time |
|------------|-------------|----------|
| Monolithic | 253         | 253      |
| Split      | 258         | 259      |

`RTX4080, CMS2018 ECAL+HCAL 8 threads x 32 events, No field`
| Kernels    | Time Master | New Time |
|------------|-------------|----------|
| Monolithic | 298         | 297      |