Refactoring the monolithic kernels for potential stepping action (#429)
This PR refactors the monolithic kernels to enable a central spot for
implementing a custom stepping actions for experiments.

As the experiment's frameworks have their own custom stepping actions
where they decide to kill unwanted tracks, it would be beneficial if
AdePT supports implementing such a stepping action in a straightforward
way.

Currently, the `survive()` function was called in multiple places in the
monolithic kernels. Therefore, this PR suggests to refactor the kernels
such that only a surviveFlag is set to true if a track survives such
that before the actual `survive` call at the end of the kernel, a
stepping action could be added.

Note that this change originally did affect the run time as it forced
some synchronization under the hood. After adding a few `__syncwarp()`
after divergent parts of the code, the run time is recovered.

For 32 threads and 64 ttbar events, the run time for master was `126.095
s` and for this PR `124.882 s`. The results are unchanged, as the RNG or
physics are not touched in this refactoring.

Note, the PR should be reviewed with `hide white space = on` as the
indentation in the gamma kernel needed to be changed, so it looks like
more changes than there actually are.

It has to be seen if this is feasible also in the split kernels, but
seems unlikely.