Fix access to HostTrackData (#430)
This PR fixes a bug in the access of the HostTrackData, so only a local
copy was modified and not the stored data.
This led to many tracks using the default constructed HostTrackData. The
fix does not change the run time. This was found during the LHCb
validation, as the joining of hits with their parents did not work:

<img width="672" height="426" alt="Screenshot from 2025-08-21 14-01-14"
src="https://github.com/user-attachments/assets/208bb180-c18b-41ce-bd7b-57ed4a7df14a"
/>

Now it does:
<img width="685" height="426" alt="Screenshot from 2025-08-21 13-59-57"
src="https://github.com/user-attachments/assets/a86aac21-a5c9-4dab-82c1-3d5499348978"
/>


- [x] test in LHCb